# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Exercise Tracker API",
    "description": "Una API para registrar usuarios y sus ejercicios",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/api/users": {
      "post": {
        "summary": "Crear un nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "juan123"
                  }
                },
                "required": ["username"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario creado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Obtener todos los usuarios",
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{_id}/exercises": {
      "post": {
        "summary": "Añadir un ejercicio a un usuario",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "example": "Correr"
                  },
                  "duration": {
                    "type": "integer",
                    "example": 30
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-06-12"
                  }
                },
                "required": ["description", "duration"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ejercicio añadido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Exercise"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{_id}/logs": {
      "get": {
        "summary": "Obtener el log de ejercicios de un usuario",
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "abc123"
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2023-01-01"
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "example": "2024-01-01"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Log de ejercicios del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Log"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "_id": {
            "type": "string"
          }
        }
      },
      "Exercise": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          },
          "_id": {
            "type": "string"
          }
        }
      },
      "Log": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "_id": {
            "type": "string"
          },
          "count": {
            "type": "integer"
          },
          "log": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string"
                },
                "duration": {
                  "type": "integer"
                },
                "date": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}